+++ python/plot3d/__init__.py
@@ -1,5 +1,7 @@
-from .face import Face, split_face, create_face_from_diagonals, find_connected_face, find_face
-from .block import Block, rotate_block, reduce_blocks
+from .face import Face
+from .facefunctions import create_face_from_diagonals, split_face,outer_face_dict_to_list, match_faces_dict_to_list, find_connected_face, find_face
+from .block import Block
+from .blockfunctions import rotate_block, reduce_blocks
 from .connectivity import get_outer_faces, find_matching_blocks, get_face_intersection, connectivity, connectivity_fast, block_connection_matrix
 from .read import read_plot3D, read_ap_nasa
 from .write import write_plot3D
