+++ python/plot3d/blockfunctions.py
@@ -1,0 +1,99 @@
+import numpy as np
+from typing import List
+from .block import Block
+
+def rotate_block(block,rotation_matrix:np.ndarray) -> Block:
+    """Rotates a block by a rotation matrix 
+
+    Args:
+        rotation_matrix (np.ndarray): 3x3 rotation matrix 
+
+    Returns:
+        Block: returns a new rotated block 
+    """
+    X = block.X.copy()
+    Y = block.Y.copy()
+    Z = block.Z.copy()
+    points = np.zeros(shape=(3,block.IMAX*block.JMAX*block.KMAX))
+    indx = 0
+    for i in range(block.IMAX):
+        for j in range(block.JMAX):
+            for k in range(block.KMAX):
+                points[0,indx] = block.X[i,j,k]
+                points[1,indx] = block.Y[i,j,k]
+                points[2,indx] = block.Z[i,j,k]
+                indx+=1
+    points_rotated = np.matmul(rotation_matrix,points)
+    indx=0
+    for i in range(block.IMAX):
+        for j in range(block.JMAX):
+            for k in range(block.KMAX):
+                X[i,j,k] = points_rotated[0,indx]
+                Y[i,j,k] = points_rotated[1,indx]
+                Z[i,j,k] = points_rotated[2,indx]
+                indx+=1
+                
+    return Block(X,Y,Z)
+                    
+
+
+def reduce_blocks(blocks:List[Block],factor:int):
+    """reduce the blocks by a factor of (factor)
+
+    Args:
+        blocks (List[Block]): list of blocks to reduce in size
+        factor (int, optional): Number of indicies to skip . Defaults to 2.
+
+    Returns:
+        [type]: [description]
+    """
+    for i in range(len(blocks)):
+        blocks[i].X = blocks[i].X[::factor,::factor,::factor]
+        blocks[i].Y = blocks[i].Y[::factor,::factor,::factor]
+        blocks[i].Z = blocks[i].Z[::factor,::factor,::factor]
+        blocks[i].IMAX,blocks[i].JMAX,blocks[i].KMAX = blocks[i].X.shape
+    return blocks
+
+def get_outer_bounds(blocks:List[Block]):
+    """Get outer bounds for a set of blocks
+
+    Args:
+        blocks (List[Block]): Blocks defining your shape
+
+    Returns:
+        (Tuple) containing: 
+
+            **xbounds** (Tuple[float,float]): xmin,xmax
+            **ybounds** (Tuple[float,float]): ymin,ymax
+            **zbounds** (Tuple[float,float]): zmin,zmax
+    """
+    xbounds = [blocks[0].X.min(),blocks[0].X.max()]
+    ybounds = [blocks[0].Y.min(),blocks[0].Y.max()]
+    zbounds = [blocks[0].Z.min(),blocks[0].Z.max()]
+    
+    for i in range(1,len(blocks)):
+        xmin = blocks[i].X.min()
+        xmax = blocks[i].X.max()
+
+        ymin = blocks[i].Y.min()
+        ymax = blocks[i].Y.max()
+
+        zmin = blocks[i].Z.min()
+        zmax = blocks[i].Z.max()
+
+        if xmin<xbounds[0]:
+            xbounds[0] = xmin
+        elif xmax>xbounds[1]:
+            xbounds[1] = xmax
+        
+        if ymin<ybounds[0]:
+            ybounds[0] = ymin
+        elif ymax>ybounds[1]:
+            ybounds[1] = ymax
+
+        if zmin<zbounds[0]:
+            zbounds[0] = zmin
+        elif zmax>zbounds[1]:
+            zbounds[1] = zmax
+    
+    return tuple(xbounds),tuple(ybounds),tuple(zbounds)
